+++ src/protocol/Channel.cc
@@ -437,7 +437,7 @@ Channel::handleChannelData(ChannelDataMsgPtr dataMsg)
     if (ev)
         ev->ChannelDataReceived(thisPtr);
     else
-        LOGE(channelId, ": Event handler required but not found");
+        LOGE(channelId, ": Event handler required but not found"); //it might be that local size socker isn't opened yet.
 }
 
 void